requirements.txt
Django==5.0
djangorestframework==3.15.0
requests==2.28.2
